backend/pyproject.toml
[project]
name = "evmbench-backend"
version = "0.1.0"
description = "Backend services for the evmbench self-hosted app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.128.0",
    "uvicorn==0.40.0",
    "pydantic-settings==2.12.0",
    "loguru==0.7.3",
    "aio-pika==9.5.8",
    "orjson==3.11.6",
    "httpx==0.28.1",
    "aiofiles==25.1.0",
    "aiodocker==0.25.0",
    "filelock==3.20.3",
    "sqlalchemy==2.0.46",
    "alembic==1.18.1",
    "asyncpg==0.31.0",
    "python-multipart==0.0.22",
    "greenlet==3.3.1",
    "kubernetes>=35.0.0",
    "pyjwt>=2.10.1",
    "pycryptodome>=3.20.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
    "ty>=0.0.14",
    "tzdata>=2025.3",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D211", "D212", "D203", "COM812", "ISC001",
    "W391", "D", "FIX001", "FIX002", "TD003",
    "S104", "INP001", "ERA001",
]
flake8-quotes.inline-quotes = "single"

[tool.ruff.format]
quote-style = "single"
line-ending = "lf"
exclude = []

[tool.ruff.lint.isort]
lines-after-imports = 2
combine-as-imports = false

[tool.ruff.lint.extend-per-file-ignores]
## Legacy paths intentionally removed from OSS release.

[tool.ty]
src.exclude = []
# uvloop imports, todo: remove this
rules.unused-ignore-comment = "ignore"
